⚡ Float -> Long
경험치의 특성과 현재 구조에선, Float보다 Long이 어올리는 것 같아 변경하였습니다.

+++ stat-api/src/main/kotlin/io/github/mulgyeolpyo/stat/StatManager.kt
@@ -9,7 +9,7 @@ package io.github.mulgyeolpyo.stat
 interface StatManager {
     /**
      * Unregisters a specific stat from this manager.
-     * If the stat's value is currently loaded in memory, it will be saved before removal.
+     * If the stat's exp is currently loaded in memory, it will be saved before removal.
      *
      * @param stat The unique identifier of the stat to unregister.
      * @throws IllegalArgumentException if the specified stat does not exist in the global stat registry.
@@ -17,17 +17,17 @@ interface StatManager {
     fun unregister(stat: String)
 
     /**
-     * Retrieves the current value of a specific stat for the managed player.
-     * If the stat value is not yet loaded into memory, it will be loaded from persistent storage.
+     * Retrieves the current exp of a specific stat for the managed player.
+     * If the stat exp is not yet loaded into memory, it will be loaded from persistent storage.
      *
      * @param stat The unique identifier of the stat to retrieve.
-     * @return The current float value of the stat.
+     * @return The current Long exp of the stat.
      * @throws IllegalArgumentException if the specified stat does not exist in the global stat registry.
      */
-    fun get(stat: String): Float
+    fun get(stat: String): Long
 
     /**
-     * Calculates and returns the level corresponding to the current value of a specific stat.
+     * Calculates and returns the level corresponding to the current exp of a specific stat.
      * The level is determined based on the stat's configuration.
      *
      * @param stat The unique identifier of the stat whose level is to be determined.
@@ -37,24 +37,24 @@ interface StatManager {
     fun level(stat: String): Int
 
     /**
-     * Sets the value of a specific stat for the managed player.
+     * Sets the exp of a specific stat for the managed player.
      * This updates the stat in memory; call [save] to persist changes.
      *
      * @param stat The unique identifier of the stat to set.
-     * @param value The new float value for the stat.
+     * @param value The new Long exp for the stat.
      * @throws IllegalArgumentException if the specified stat does not exist in the global stat registry.
      */
     fun set(
         stat: String,
-        value: Float,
+        value: Long,
     )
 
     /**
-     * Sets the value of a specific stat for the managed player, converting an integer to a float.
+     * Sets the exp of a specific stat for the managed player, converting an integer to a Long.
      * This updates the stat in memory; call [save] to persist changes.
      *
      * @param stat The unique identifier of the stat to set.
-     * @param value The new integer value for the stat, which will be converted to float.
+     * @param value The new integer exp for the stat, which will be converted to Long.
      * @throws IllegalArgumentException if the specified stat does not exist in the global stat registry.
      */
     fun set(
@@ -63,75 +63,75 @@ interface StatManager {
     )
 
     /**
-     * Increases the value of a specific stat for the managed player by a float amount.
+     * Increases the exp of a specific stat for the managed player by a Long amount.
      * This updates the stat in memory; call [save] to persist changes.
      *
      * @param stat The unique identifier of the stat to increment.
-     * @param value The float amount to add to the current stat value.
+     * @param value The Long amount to add to the current stat exp.
      * @throws IllegalArgumentException if the specified stat does not exist in the global stat registry.
      */
     fun increment(
         stat: String,
-        value: Float,
-    ): Float
+        value: Long,
+    ): Long
 
     /**
-     * Increases the value of a specific stat for the managed player by an integer amount.
+     * Increases the exp of a specific stat for the managed player by an integer amount.
      * This updates the stat in memory; call [save] to persist changes.
      *
      * @param stat The unique identifier of the stat to increment.
-     * @param value The integer amount to add to the current stat value, which will be converted to float.
+     * @param value The integer amount to add to the current stat exp, which will be converted to Long.
      * @throws IllegalArgumentException if the specified stat does not exist in the global stat registry.
      */
     fun increment(
         stat: String,
         value: Int,
-    ): Float
+    ): Long
 
     /**
-     * Decreases the value of a specific stat for the managed player by a float amount.
+     * Decreases the exp of a specific stat for the managed player by a Long amount.
      * This updates the stat in memory; call [save] to persist changes.
      *
      * @param stat The unique identifier of the stat to decrement.
-     * @param value The float amount to subtract from the current stat value.
+     * @param value The Long amount to subtract from the current stat exp.
      * @throws IllegalArgumentException if the specified stat does not exist in the global stat registry.
      */
     fun decrement(
         stat: String,
-        value: Float,
-    ): Float
+        value: Long,
+    ): Long
 
     /**
-     * Decreases the value of a specific stat for the managed player by an integer amount.
+     * Decreases the exp of a specific stat for the managed player by an integer amount.
      * This updates the stat in memory; call [save] to persist changes.
      *
      * @param stat The unique identifier of the stat to decrement.
-     * @param value The integer amount to subtract from the current stat value, which will be converted to float.
+     * @param value The integer amount to subtract from the current stat exp, which will be converted to Long.
      * @throws IllegalArgumentException if the specified stat does not exist in the global stat registry.
      */
     fun decrement(
         stat: String,
         value: Int,
-    ): Float
+    ): Long
 
     /**
-     * Loads the value of a specific stat for the managed player from persistent storage into memory.
-     * If the player's data for this stat is not found, it defaults to the stat's configured random initial value.
+     * Loads the exp of a specific stat for the managed player from persistent storage into memory.
+     * If the player's data for this stat is not found, it defaults to the stat's configured random initial exp.
      *
      * @param stat The unique identifier of the stat to load.
-     * @return The loaded float value of the stat.
+     * @return The loaded Long exp of the stat.
      * @throws IllegalArgumentException if the specified stat does not exist in the global stat registry.
      */
-    fun load(stat: String): Float
+    fun load(stat: String): Long
 
     /**
-     * Loads the values of all registered stats for the managed player from persistent storage into memory.
+     * Loads the exps of all registered stats for the managed player from persistent storage into memory.
      * This typically iterates through all stats known to the global manager and calls [load] for each.
      */
     fun load()
 
     /**
-     * Saves the current in-memory value of a specific stat for the managed player to persistent storage.
+     * Saves the current in-memory exp of a specific stat for the managed player to persistent storage.
      *
      * @param stat The unique identifier of the stat to save.
      * @throws IllegalArgumentException if the specified stat does not exist in the global stat registry.
@@ -139,8 +139,8 @@ interface StatManager {
     fun save(stat: String)
 
     /**
-     * Saves the current in-memory values of all managed stats for the player to persistent storage.
-     * This typically iterates through all currently known stat values and calls [save] for each.
+     * Saves the current in-memory exps of all managed stats for the player to persistent storage.
+     * This typically iterates through all currently known stat exps and calls [save] for each.
      */
     fun save()
 }
